+++ policyengine-client/src/policyengine/pages/household/household.jsx
@@ -6,21 +6,9 @@ import { VariableControlPane } from "./inputPane";
 import AccountingTable from "./accountingTable";
 import EarningsChartsPane from "./earningsCharts";
 import { OverviewHolder, PolicyOverview } from "../policy/overview";
-import { Divider, Collapse } from "antd";
+import { Divider } from "antd";
 import NavigationButton from "../../general/navigationButton";
-import { ArrowLeftOutlined, ExclamationCircleOutlined } from "@ant-design/icons";
-
-const { Panel } = Collapse;
-
-function HouseholdResultsCaveats() {
-	return (
-		<Collapse style={{marginTop: 15}} defaultActiveKey={["0"]} ghost>
-			<Panel header={<><ExclamationCircleOutlined />  &nbsp; &nbsp;Disclaimer</>} key="1">
-				<p>Results are calculated using the OpenFisca-US tax-benefit microsimulation model, may not constitute exact benefit entitlement. See the <a href="https://github.com/PSLmodels/openfisca-uk">repository</a> for more information.</p>
-			</Panel>
-		</Collapse>
-	);
-}
+import { ArrowLeftOutlined } from "@ant-design/icons";
 
 export class Household extends React.Component {
     static contextType = CountryContext;
@@ -57,15 +45,9 @@ export class Household extends React.Component {
                 variables={this.getVariables()} 
             />;
         } else if(this.state.selected === "results") {
-            middlePane = <>
-                <AccountingTable />
-                <HouseholdResultsCaveats />
-            </>;
+            middlePane = <AccountingTable />;
         } else if(this.state.selected === "earnings") {
-            middlePane = <>
-                <AccountingTable />
-                <EarningsChartsPane />
-            </>;
+            middlePane = <EarningsChartsPane />;
         }
         return <>
             <Row>
